coca_cola dataset using forcasting algorithm
PROBLEM STATEMENT
The objective of this project is to build a forecasting model to predict the sales of Coca-Cola. The dataset contains historical values of the sales, recorded quarterly over a certain period. The aim is to use various forecasting algorithms, such as ARIMA, SARIMA, LSTM, etc., to predict the future values of the sales.

The primary goal of this project is to determine which forecasting algorithm performs best for the given Coca-Cola sales dataset. The accuracy of the forecasting models will be evaluated using standard metrics such as mean absolute error, mean squared error, and root mean squared error. Additionally, the suitability of each algorithm to the specific time series dataset and the interpretability of the results will also be considered.

The final outcome of this project will be the selection of the most accurate and appropriate forecasting algorithm for the given Coca-Cola sales dataset. This will enable the company to forecast future sales with greater accuracy and confidence, allowing for better planning and decision-making.